requirements.txt
ai2thor==5.0.0
allenact==0.5.4
allenact_plugins==0.5.4
apex==0.9.10dev
backend==0.2.4.1
canonicaljson==2.0.0
git+https://github.com/openai/CLIP.git
compress_pickle==2.1.0
faiss_cpu==1.8.0.post1
FastSAM==0.0.1.dev0
gym==0.17.3
ipdb==0.13.13
lru_dict==1.3.0
matplotlib
moviepy==1.0.3
mmcv_full==1.7.1
mmdet==3.3.0
mmsegmentation==0.27.0
natsort==8.4.0
nets==0.0.3.1
networkx==3.2.1
numpy-quaternion
omegaconf==2.3.0
onnxruntime==1.19.2
open3d==0.18.0
opencv_python==4.10.0.84
opencv_python_headless==4.10.0.84
open_clip_torch
packaging==24.2
Pillow==11.0.0
plyfile==1.1
pycocotools==2.0.8
PyMCubes==0.1.4
pymeshlab==2023.12.post2
pynput==1.7.7
rich==13.9.4
segment_anything
scikit_learn==1.6.0
scipy
setproctitle==1.3.3
scikit-image
scikit-fmm
stringcase==1.2.0
submitit==1.5.2
supervision==0.25.1
tqdm==4.65.2
tensorboard
-e ./gaussian_splatting/submodules/simple-knn
-e ./gaussian_splatting/submodules/diff-gaussian-rasterization
-e ./gaussian_splatting/submodules/langsplat-rasterization
-e ./gaussian_splatting/submodules/segment-anything-langsplat
